refactor(search): reorganize app search files
